requirements.txt
Flask~=3.0.3
Flask-SQLAlchemy
psycopg2-binary
requests~=2.32.3
Flask-Cors~=5.0.0
SQLAlchemy~=2.0.35
Flask-Bcrypt~=1.0.1
WTForms~=3.1.2
Flask-Migrate~=4.0.7
Flask-JWT-Extended~=4.6.0
alembic~=1.13.3
Flask-Login~=0.6.3